+++ src/domain/CreateEventStreamRecord.ts
